New method of scanning for packages, this checks the installed package list against the request
Rather than assuming the presence of a "packages-up-to-date" file indicates the packages are actually
up to date, we compute a SHA-1 hash of the installed package list and compare it against the requested
list. Only packages in the emacs-init-script's package-list will be added to the hash.

As a result of this, editing the package-list on one computer should automatically invoke updates on
other computers (after they pull down the changes) without a human noticing that the packages are out
of date and deleting the packages-up-to-date file